WIP: go back to center-of-mass centroiding
If the two speckles are bright enough, then the intensity distribution
in their cores should be close enough that the the difference of their
center-of-mass coordinates should be a good estimator of the true
speckle separation.  In simulation, this reduces mean-squared error
by about a factor of 2 compared to the cross-correlation method.