RTFM on subclassing astropy.units.Quantity
https://docs.astropy.org/en/stable/units/quantity.html#subclassing-quantity

```
Subclassing Quantity
To subclass Quantity, you generally proceed as you would when subclassing ndarray (i.e., you typically need to override __new__, rather than __init__, and use the numpy.ndarray.__array_finalize__ method to update attributes). For details, see the NumPy documentation on subclassing. To get a sense of what is involved, have a look at Quantity itself, where, for example, the astropy.units.Quantity.__array_finalize__ method is used to pass on the unit, at Angle, where strings are parsed as angles in the astropy.coordinates.Angle.__new__ method and at Longitude, where the astropy.coordinates.Longitude.__array_finalize__ method is used to pass on the angle at which longitudes wrap.

Another method that is meant to be overridden by subclasses, one specific to Quantity, is astropy.units.Quantity.__quantity_subclass__. This is called to decide which type of subclass to return, based on the unit of the quantity that is to be created. It is used, e.g., in Angle to return a Quantity if a calculation returns a unit other than an angular one. The implementation of this is via SpecificTypeQuantity, which more generally allows one to construct Quantity subclasses that have methods that are useful only for a specific physical type.
```

Signed-off-by: James Noss <jnoss@stsci.edu>